requirements.txt
cozepy~=0.19.0
python-dotenv~=0.21.0
pathlib~=1.0.1